Move process execution to using 'spawn' intead of 'fork'
Added wrapper class for launching new processes via the multiprocessing
'spawn' method. This was necessary as before the target method I was
using contained decorator wrapped functions which the pickler didn't
like.

I also had to add a layer of 'dill' ing so functions, and arguments
could be properly passed to the new process.